nitorassessmentportal-fix-test_page/nitoronlinetestportal/requirements.txt
asgiref==3.6.0
cfgv==3.3.1
coverage==7.2.3
distlib==0.3.6
Django==3.1.7
django-extensions==3.1.5
django-rest-framework==0.1.0
django-webpack-loader==0.7.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
filelock==3.12.0
identify==2.5.24
iniconfig==2.0.0
isort==5.12.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
psycopg2==2.9.5
pydot==1.4.2
pydotplus==2.0.2
pyparsing==3.0.9
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
virtualenv==20.23.0
